requirements.txt
# Sabra Device Backup - Python Dependencies
# Django Web Framework
Django>=5.0,<6.0
gunicorn>=21.0.0

# Database
psycopg2-binary>=2.9.9

# Task Queue
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Network Automation
netmiko>=4.3.0
paramiko>=3.4.0
textfsm>=1.1.3
ntc-templates>=4.0.0

# Security & Encryption
django-fernet-fields-v2>=0.9
cryptography>=41.0.0
django-axes>=6.3.0  # Brute force protection

# Forms & Validation
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2

# Utilities
python-dateutil>=2.8.2
django-filter>=23.5
Pillow>=10.0.0

# Development & Testing
django-debug-toolbar>=4.2.0

# Diff & Text Processing
diff-match-patch>=20230430

# Environment (for development only, production uses encrypted DB storage)
python-dotenv>=1.0.0
